Update platforms command
